LASB-843|added unit test for childweightings and updated existing tests

+++ maat-court-data-api/src/main/java/gov/uk/courtdata/model/assessment/ChildWeightings.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ChildWeightings {
-    private Long childWeightingId;
+    private Integer childWeightingId;
     private Integer noOfChildren;
 }
